+++ modules/discovery/aliases.go
@@ -1,6 +1,0 @@
-package discovery
-
-import "github.com/brainlet/brainkit/internal/types"
-
-type Config = types.DiscoveryConfig
-type PeerConfig = types.PeerConfig